% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{read_integer}
\alias{read_integer}
\title{Read Object as Integer}
\usage{
read_integer(x)
}
\arguments{
\item{x}{an object.}
}
\value{
An integer.
}
\description{
Read any R object as an integer. This function dereferences the pointer to
    the R object and reads the value at the address as an integer. Useful for
    converting 32-bit (4-byte) raw vectors to the equivalent integer value.
}
\examples{
# SHAKE256 hash to integer:
read_integer(sha3("secret base", size = 32L, convert = FALSE))

}
