% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\docType{package}
\encoding{UTF-8}
\name{secretbase-package}
\alias{secretbase}
\alias{secretbase-package}
\title{secretbase: Cryptographic Hash and Extendable-Output Functions}
\description{
Fast and memory-efficient implementations of the SHA-256, SHA-3 cryptographic
    hash functions, SHAKE256 extendable-output function (XOF), and 'SipHash'
    pseudo-random function. \cr\cr The SHA-3 cryptographic hash functions are
    SHA3-224, SHA3-256, SHA3-384 and SHA3-512, each an instance of the Keccak
    algorithm. SHAKE256 is one of the two XOFs of the SHA-3 family, along
    with SHAKE128 (not implemented).
}
\references{
The SHA-3 Secure Hash Standard was published by the National
    Institute of Standards and Technology (NIST) in 2015 at
    \doi{doi:10.6028/NIST.FIPS.202}.
    
    The SHA-256 Secure Hash Standard was published by NIST in 2002 at
    \url{https://csrc.nist.gov/publications/fips/fips180-2/fips180-2.pdf}.
    
    \pkg{secretbase} adapts the SHA-256 and SHA-3 implementations by the
    'Mbed TLS Contributors' at \url{https://github.com/Mbed-TLS/mbedtls} 
    under the 'Mbed TLS' Trusted Firmware Project
    \url{https://www.trustedfirmware.org/projects/mbed-tls/}.
    
    The SipHash family of cryptographically-strong pseudorandom functions
    (PRFs) are described in 'SipHash: a fast short-input PRF', Jean-Philippe
    Aumasson and Daniel J. Bernstein, Paper 2012/351, 2012, Cryptology ePrint
    Archive at \url{https://ia.cr/2012/351}.
    
    \pkg{secretbase} adapts the SipHash streaming implementation by Daniele
    Nicolodi, David Rheinsberg and Tom Gundersen at
    \url{https://github.com/c-util/c-siphash}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shikokuchuo.net/secretbase/}
  \item \url{https://github.com/shikokuchuo/secretbase/}
  \item Report bugs at \url{https://github.com/shikokuchuo/secretbase/issues}
}

}
\author{
Charlie Gao \email{charlie.gao@shikokuchuo.net}
    (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
}
